Enhance test
