+++ test/utils/MediaStream.js
@@ -1,0 +1,6 @@
+class MediaStream {
+  id = ''
+  active = true
+}
+
+export default MediaStream;
