+++ test/utils/NavigatorUserMediaError.js
@@ -1,0 +1,7 @@
+class NavigatorUserMediaError extends Error {
+  name = 'PermissionDeniedError'
+  constraintName = ''
+  message = ''
+}
+
+export default NavigatorUserMediaError;
